+++ CHANGELOG.md
@@ -14,6 +14,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Removed
 
+## [1.4.0]
+
+### Added
+
+- Test infrastructure, setup, and new tests for pulling dataset, reporting org
+  lists from SuiteCRM.
+- Code to pull the dataset and reporting org lists from SuiteCRM.
+
+### Fixed
+
+- Deleted datasets are only deleteed from Azure blob storage if there is a
+  cached copy stored.
+
 ## [1.3.4]
 
 ### Added